% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_base.R
\name{bart}
\alias{bart}
\title{helper of bart class}
\usage{
bart(
  name,
  genome,
  gene_data,
  region_data,
  gene_mode_param = list(binsize = 1000),
  region_mode_param = list(binsize = 50, scorecol = 5),
  bimodal_mode_param = list(binsize = 50, weights = c(1, 1))
)
}
\arguments{
\item{name}{name of the input data, it could be a sample name, a cluster id and so on}

\item{genome}{"mm10" or "hg38"}

\item{gene_data}{a vector of gene symbols, check the internal data B_cell_gene as an example}

\item{region_data}{a dataframe that follows a BED6 format, check the internal data B_cell_region as an example}

\item{gene_mode_param}{a list of costumized arguments for gene mode}

\item{region_mode_param}{a list of costumized arguments for region mode}

\item{bimodal_mode_param}{a list of arguments for bimodal mode. Currently
supports \code{binsize} and \code{weights}. \code{weights} must be a non-negative numeric
vector of length 2 giving the relative weight of RNA and ATAC ranks;
it is normalized internally to sum to 1. Default is \code{c(1, 1)}. A zero
weight disables that modality in the rank aggregation.}
}
\value{
A bart object
}
\description{
This function creates a bart object with specified data.
}
