% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\docType{class}
\name{bartsc-class}
\alias{bartsc-class}
\title{An S4 class to store raw data, intermediate data and outcomes of BARTsc analysis}
\description{
An S4 class to store raw data, intermediate data and outcomes of BARTsc analysis
}
\section{Slots}{

\describe{
\item{\code{meta,}}{param, data, Args, intermediate, result}
}}

