% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{bartsc}
\alias{bartsc}
\title{Create a bartsc object.}
\usage{
bartsc(
  name,
  genome,
  label,
  cell_types_used = NULL,
  RNA_cnt_matrix = NULL,
  RNA_norm_matrix = NULL,
  ATAC_cnt_matrix = NULL,
  ATAC_norm_matrix = NULL,
  peaks = NULL,
  active_mod = NULL,
  gene_mode_param = list(binsize = 1000),
  region_mode_param = list(binsize = 50, scorecol = 5),
  bimodal_mode_param = list(binsize = 50, weights = c(1, 1))
)
}
\arguments{
\item{name}{Name to use for the object.}

\item{genome}{Genome to use, hg38 or mm10.}

\item{label}{A vector of cell types, with cell ids as names.}

\item{cell_types_used}{A vector of cell types to use. cell types must exist
in \code{label}.}

\item{RNA_cnt_matrix}{A gene by cell read count matrix}

\item{RNA_norm_matrix}{A gene by cell normalized gene expression matrix.
Must be a \code{dgCMatrix} object.}

\item{ATAC_cnt_matrix}{A peak by cell count matrix. Must be a \code{dgCMatrix}
object.}

\item{ATAC_norm_matrix}{A peak by cell normalized chromatin
accessibility matrix. Must be a \code{dgCMatrix} object.}

\item{peaks}{A dataframe of input peaks.}

\item{active_mod}{Active modality. One of "RNA", "ATAC", or "bimodal".
Required when constructing an object without count matrices (e.g. for
custom feature analysis). If NULL, inferred from provided matrices.}

\item{gene_mode_param}{A list of parameters for gene mode.}

\item{region_mode_param}{A list of parameters list for region mode.}

\item{bimodal_mode_param}{A list of parameters for bimodal mode. Currently
supports \code{binsize} and \code{weights}. \code{weights} must be a non-negative numeric
vector of length 2 giving the relative weight of RNA and ATAC ranks;
it is normalized internally to sum to 1. Default is \code{c(1, 1)}. A zero
weight disables that modality in the rank aggregation.}
}
\value{
a bartsc object
}
\description{
bartsc object is used to store input data, processed data and results of
BARTsc analyses. With required input data, this function creates a bartsc
object of scRNA-seq, scATAC-seq or the bimodal sequencing of transcriptiome
and chromatin accessibility (single-cell multiomics).
}
