% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{calc_crossCT_auc_ATAC}
\alias{calc_crossCT_auc_ATAC}
\title{calculate aucs for cross-cell-type analysis of ATAC}
\usage{
calc_crossCT_auc_ATAC(object)
}
\arguments{
\item{object}{a bart object}
}
\value{
a bartsc object
}
\description{
calculate aucs for cross-cell-type analysis of ATAC
}
