% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{crossCT_test}
\alias{crossCT_test}
\title{create bart objects and run BART for cross-cell-type test}
\usage{
crossCT_test(object, mod)
}
\arguments{
\item{object}{a bart object}

\item{mod}{one of "RNA", "ATAC" and "bimodal"}
}
\value{
a bartsc object
}
\description{
create bart objects and run BART for cross-cell-type test
}
