% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{crossCT_test_ATAC}
\alias{crossCT_test_ATAC}
\title{create bart objects and run BART for cross-cell-type test}
\usage{
crossCT_test_ATAC(object)
}
\arguments{
\item{object}{a bart object}
}
\value{
a bartsc object
}
\description{
create bart objects and run BART for cross-cell-type test
}
