% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{deviation_heatmap}
\alias{deviation_heatmap}
\title{Heatplot visualization of cross-cell type analysis, colored by deviation ratio}
\usage{
deviation_heatmap(
  object,
  mod,
  tf,
  cell_types_used = NULL,
  value.display = TRUE,
  decimal = 2,
  fontsize = 18,
  tile_fontsize = 7
)
}
\arguments{
\item{object}{A bartsc object.}

\item{mod}{One of "RNA", "ATAC" and "bimodal".}

\item{tf}{The transcription factor or chromatin regulator to inspect.}

\item{cell_types_used}{Cell types to use, must be a subset of the object@meta$cell_types_used.}

\item{value.display}{Whether to display the deviation ratio in heatmap tiles.}

\item{decimal}{The number of decimal places to retain for the displayed values. Default it 2.}

\item{fontsize}{size of font}

\item{tile_fontsize}{font size of displayed values in heatmap tiles.}
}
\value{
A plot object.
}
\description{
Heatplot visualization of cross-cell type analysis, colored by deviation ratio
}
