% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dot_plot}
\alias{dot_plot}
\title{Dot plot visualization of cross-cell type analysis}
\usage{
dot_plot(
  object,
  mod,
  tf,
  cell_types_used = NULL,
  max_dot_size = 20,
  fontsize = 18
)
}
\arguments{
\item{object}{A bartsc object.}

\item{mod}{One of "RNA", "ATAC" and "bimodal".}

\item{tf}{The transcription factor or chromatin regulator to inspect.}

\item{cell_types_used}{Cell types to use, must be a subset of the object@meta$cell_types_used.}

\item{max_dot_size}{Maximum size of dots.}

\item{fontsize}{Size of font.}
}
\value{
A plot object.
}
\description{
Dot plot visualization of cross-cell type analysis
}
