% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{find_key_regulators}
\alias{find_key_regulators}
\title{Make a 3d scatter plot for key regulators identification}
\usage{
find_key_regulators(object, mod, min.N.profile = 3)
}
\arguments{
\item{object}{A bartsc object.}

\item{mod}{One of "RNA", "ATAC" and "bimodal".}

\item{min.N.profile}{Only transcription regulators with at least certain
number of binding profiles in database are considered. Default is 3.}
}
\value{
A bartsc object.
}
\description{
BARTsc identifies key regulators for a given cell types by comprehensively
considering three factors: 1) Contribution to signature expression, 2) rela-
tive activity across all the cell types and 3) the uniqueness of a given
transcription regulator. These three factors are quantified by signature
score, MDR (mean deviation ratio) and dMDR (differential MDR) respectively.
Final rank is determined by the mean rank of above indices. Final pvalues
are calculated based on Irvin-Hall distribution. To garantee the robustness
of final output, only transcription regulators with at least 3 binding pro-
files in database are considered by default.
}
