% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc_find_differential.R
\name{find_signature_genes}
\alias{find_signature_genes}
\title{Find cell type signature genes}
\usage{
find_signature_genes(
  object,
  min.pct = 0.1,
  min.diff.pct = -Inf,
  log2fc.thr = 1,
  pval.thr = NULL,
  padj.thr = 0.05,
  auc.thr = 0,
  max.cells.per.ident = Inf
)
}
\arguments{
\item{object}{A bartsc object.}

\item{min.pct}{Only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the
function by not testing genes that are very infrequently expressed. Default
is 0.1.}

\item{min.diff.pct}{Only test genes that show a minimum difference in the
fraction of detection between the two groups. Set to -Inf by default.}

\item{log2fc.thr}{Threshold of log2(foldchange).}

\item{pval.thr}{Threshold of p value.}

\item{padj.thr}{Threshold of adjusted p value.}

\item{auc.thr}{Threshold of AUC calculated by presto.}

\item{max.cells.per.ident}{Down sample each identity class to a max number.
Not activated by default (set to Inf).}
}
\value{
object A bartsc object.
}
\description{
Find cell type signature genes by comparing the target cell type with other
cells. Wilcoxon rank sum test and Benjamini-Hochberg FDR adjustment.
}
