% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bart_base.R
\name{generateArgs}
\alias{generateArgs}
\title{Create a Args Namespace}
\usage{
generateArgs(object, subcommand)
}
\arguments{
\item{object}{a Bart object}

\item{subcommand}{"geneset" or "region"}
}
\value{
a python Namespace
}
\description{
This function create a python Namespace that holds BART2 arguments.
}
