% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_utils.R
\name{get_bart_auc}
\alias{get_bart_auc}
\title{return BART AUC values}
\usage{
get_bart_auc(object, subcommand)
}
\arguments{
\item{object}{a bart object}

\item{subcommand}{"geneset" or "region" or "bimodal"}
}
\value{
a dataframe of AUC values
}
\description{
This function return the ROC AUC calculate by BART
}
