% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_utils.R
\name{get_bart_result}
\alias{get_bart_result}
\title{return BART result}
\usage{
get_bart_result(object, subcommand)
}
\arguments{
\item{object}{a bart object}

\item{subcommand}{"geneset" or "region" or "bimodal"}
}
\value{
a dataframe of bart result
}
\description{
This function return the result of BART
}
