% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc_utils.R
\name{get_result}
\alias{get_result}
\title{Get the analysis results for cell type signature analysis or cross-cell-
type analysis}
\usage{
get_result(object, analysis = "cell type signature", mod)
}
\arguments{
\item{object}{A bartsc object.}

\item{analysis}{name of analysis to check, one of "cell type signature",
"cross-cell-type", "Key regs ident"}

\item{mod}{one of "RNA", "ATAC" and "bimodal"}
}
\description{
Get the analysis results for cell type signature analysis or cross-cell-
type analysis
}
