% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bart_utils.R
\name{getResult}
\alias{getResult}
\title{return BART result}
\usage{
getResult(object, subcommand)
}
\arguments{
\item{object}{a Bart object}

\item{subcommand}{"geneset" or "region" or "bimodal"}
}
\value{
a dataframe of bart result
}
\description{
This function return the result of BART
}
