% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_base.R
\name{identifyEnhancer}
\alias{identifyEnhancer}
\title{Identify active enhancers}
\usage{
identifyEnhancer(object)
}
\arguments{
\item{object}{a bart object}
}
\value{
a bart object
}
\description{
This function generate a cis-regulatory profile using the linear combination of H3K27ac profiles predicted by Marge.
}
