% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{initialize}
\alias{initialize}
\title{initialize the bart2 python package}
\usage{
initialize(condaenv = NULL, virtualenv = NULL, python_ver = "3.9")
}
\arguments{
\item{condaenv}{name of conda environment that bart2 is installed}

\item{python_ver}{python version to use, default is "3.9"}
}
\value{
no return.
}
\description{
This function install bart2, download library and rewrite bart2 config
}
\details{
This function
}
