% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{initiate}
\alias{initiate}
\title{Initiate the bart2 python package}
\usage{
initiate(condaenv = NULL, virtualenv = NULL)
}
\arguments{
\item{condaenv}{name of conda environment that bart2 is installed}
}
\value{
no return.
}
\description{
This function install bart2, download library and rewrite bart2 config
}
\details{
This function
}
