% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{key_regulator_scatter}
\alias{key_regulator_scatter}
\title{Make a 3d scatter plot for key regulators identification}
\usage{
key_regulator_scatter(
  object,
  mod,
  cell_type,
  tfs_labeled = NULL,
  pval.thr = 0.05
)
}
\arguments{
\item{object}{A bartsc object.}

\item{mod}{One of "RNA", "ATAC" and "bimodal".}

\item{cell_type}{Visualize the selected cell type, must be one of the cell
type in object@meta$cell_types_used.}

\item{tfs_labeled}{Transcription regulators to be labelled on the plot.}

\item{pval.thr}{Threshold of the final p value.}
}
\value{
A plot.
}
\description{
BARTsc identifies key regulators for a given cell types by comprehensively
considering three factors: 1) Contribution to signature expression, 2) rela-
tive activity across all the cell types and 3) the uniqueness of a given
transcription regulator. These three factors are quantified by signature
score, MDR (mean deviation ratio) and dMDR (differential MDR) respectively.
key_regulator_scatter() make a 3d scatter plot accordingly and color key
regulators passed p value threshold.
}
