% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{load_bart2}
\alias{load_bart2}
\title{load installed bart2 and modules of it}
\usage{
load_bart2(condaenv = NULL, virtualenv = "scbart_env")
}
\arguments{
\item{condaenv}{conda environemnt name}

\item{virtualenv}{virtual envrionment name}
}
\value{
no return
}
\description{
load installed bart2 and modules of it
}
