% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_base.R
\name{mapRegionScore}
\alias{mapRegionScore}
\title{Map region scores to UDHS sites}
\usage{
mapRegionScore(object)
}
\arguments{
\item{object}{a bart object}
}
\value{
a bart object
}
\description{
This function Map region scores to UDHS sites to generate a region-derived cis-regulatory profile
}
