% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{normalize_ATAC}
\alias{normalize_ATAC}
\title{Normalize the scATAC-seq count matrix in a bartsc object.}
\usage{
normalize_ATAC(object, method.use = 1, scale.factor = 10000)
}
\arguments{
\item{object}{A bartsc object.}

\item{scale.factor}{Which scale factor to use. Default is 10000.}

\item{method}{Which method to use for calculation. To check details,
use help(RunTFIDF).}
}
\value{
A bartsc object.
}
\description{
Use TF-IDF to normalize scATAC-seq data. scale factor is set as 1e4.
}
