% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{normalize_RNA}
\alias{normalize_RNA}
\title{Normalize the scRNA-seq count matrix in a bartsc object.}
\usage{
normalize_RNA(object)
}
\arguments{
\item{object}{A bartsc object.}
}
\value{
A bartsc object.
}
\description{
Feature counts for each cell are divided by the total
counts for that cell and multiplied by 10,000. This
is then natural-log transformed using ‘log1p’.
}
