% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{run_bart_bimodal}
\alias{run_bart_bimodal}
\title{run BART bimodal mode}
\usage{
run_bart_bimodal(
  object,
  name,
  genome,
  symbol_list = NULL,
  region_list = NULL,
  gene_mode_param = list(binsize = 1000),
  region_mode_param = list(binsize = 50, scorecol = 5),
  bimodal_mode_param = list(binsize = 50, weights = c(1, 1)),
  reserve_interm = FALSE,
  return_null = FALSE
)
}
\arguments{
\item{object}{a BART object with valid geneset data}

\item{name}{name of the input data, it could be a sample name, a cluster id
and so on}

\item{genome}{"mm10" or "hg38"}

\item{symbol_list}{a vector of gene symbols, check the internal data
B_cell_gene as an example}

\item{region_list}{a dataframe that follows a BED6 format, check the
internal data B_cell_region as an example}

\item{gene_mode_param}{a list of costumized arguments for gene mode}

\item{bimodal_mode_param}{a list of arguments for bimodal mode. Currently
supports \code{binsize} and \code{weights}. \code{weights} must be a positive numeric
vector of length 2 giving the relative weight of RNA and ATAC ranks;
it is normalized internally to sum to 1. Default is \code{c(1, 1)}.}
}
\value{
A bart object
}
\description{
This function run BART bimodal mode. Input could be a pre-created bart
object, or be set by specifying name, genome, symbol_list and region_list
}
