% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{run_bart_gene_set}
\alias{run_bart_gene_set}
\title{run BART geneset mode}
\usage{
run_bart_gene_set(
  object,
  name,
  genome,
  symbol_list = NULL,
  gene_mode_param = list(binsize = 1000),
  reserve_interm = FALSE,
  return_null = FALSE
)
}
\arguments{
\item{object}{a BART object with valid geneset data}

\item{name}{name of the input data, it could be a sample name, a cluster id
and so on}

\item{genome}{"mm10" or "hg38"}

\item{symbol_list}{a vector of gene symbols, check the internal data
B_cell_gene as an example}

\item{gene_mode_param}{a list of costumized arguments for gene mode}
}
\value{
A bart object
}
\description{
This function run BART geneset mode. Input could be a pre-created bart
object, or be set by specifying name, genome and symbol_list
}
