% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{run_bart_region}
\alias{run_bart_region}
\title{run BART region mode}
\usage{
run_bart_region(
  object,
  name,
  genome,
  region_list = NULL,
  region_mode_param = list(binsize = 50, scorecol = 5),
  reserve_interm = FALSE,
  return_null = FALSE
)
}
\arguments{
\item{object}{a BART object with valid region data}

\item{name}{name of the input data, it could be a sample name, a cluster id
and so on}

\item{genome}{"mm10" or "hg38"}

\item{region_list}{a dataframe that follows a BED6 format, check the
internal data B_cell_region as an example}

\item{region_mode_param}{a list of costumized arguments for region mode}
}
\value{
A bart object
}
\description{
This function run BART region mode. Input could be a pre-created bart
object, or be set by specifying name, genome and region_list
}
