% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{run_signature_ATAC}
\alias{run_signature_ATAC}
\title{Run cell type signature analysis on ATAC assay}
\usage{
run_signature_ATAC(object)
}
\arguments{
\item{object}{A bartsc object.}
}
\description{
Run cell type signature analysis on ATAC assay
}
