% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{run_signature_bimodal}
\alias{run_signature_bimodal}
\title{Run cell type signature analysis on RNA assay and ATAC assay with bimodal integration method}
\usage{
run_signature_bimodal(object)
}
\arguments{
\item{object}{A bartsc object.}
}
\description{
Run cell type signature analysis on RNA assay and ATAC assay with bimodal integration method
}
