% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsc.R
\name{run_signature_RNA}
\alias{run_signature_RNA}
\title{Run cell type signature analysis on RNA assay}
\usage{
run_signature_RNA(object)
}
\arguments{
\item{object}{A bartsc object.}
}
\description{
Run cell type signature analysis on RNA assay
}
