% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{runBartRegion}
\alias{runBartRegion}
\title{run BART region mode}
\usage{
runBartRegion(
  object,
  name,
  genome,
  region_list = NULL,
  gene_mode_param = list(binsize = 1000)
)
}
\arguments{
\item{object}{a BART object with valid region data}

\item{name}{name of the input data, it could be a sample name, a cluster id and so on}

\item{genome}{"mm10" or "hg38"}

\item{region_list}{a dataframe that follows a BED6 format, check the internal data B_cell_region as an example}

\item{region_mode_param}{a list of costumized arguments for region mode}
}
\value{
A Bart object
}
\description{
This function run BART region mode. Input could be a pre-created Bart
object, or be set by specifying name, genome and region_list
}
\examples{
bart_obj <- Bart("B_cell_region", "hg38", region_data = B_cell_region)
bart_obj <- runBartBartRegion(bart_obj)

bart_obj <- runBartBartRegion("B_cell_region", "hg38", B_cell_region)
}
