% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_base.R
\name{runMarge}
\alias{runMarge}
\title{Run Marge on geneset input}
\usage{
runMarge(object)
}
\arguments{
\item{object}{a bart object}
}
\value{
a bart object
}
\description{
This function generate a Marge model based on input gene list
}
