% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bart_base.R
\name{Bart}
\alias{Bart}
\title{helper of Bart class}
\usage{
Bart(
  name,
  genome,
  gene_data = NULL,
  region_data = NULL,
  gene_mode_param = list(binsize = 1000),
  region_mode_param = list(binsize = 50, scorecol = 5)
)
}
\arguments{
\item{name}{name of the input data, it could be a sample name, a cluster id and so on}

\item{genome}{"mm10" or "hg38"}

\item{gene_data}{a vector of gene symbols, check the internal data B_cell_gene as an example}

\item{region_data}{a dataframe that follows a BED6 format, check the internal data B_cell_region as an example}

\item{gene_mode_param}{a list of costumized arguments for gene mode}

\item{region_mode_param}{a list of costumized arguments for region mode}
}
\value{
A Bart object
}
\description{
This function creates a Bart object with specified data.
}
\examples{
bart_obj <- Bart("B_cell_gene", "hg38", gene_data = B_cell_gene)
bart_obj <- Bart("B_cell_region", "hg38", region_data = B_cell_region)
bart_obj <- Bart("B_cell_gene", "hg38",
    gene_data = B_cell_gene,
    region_data = B_cell_region
)
}
