% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodal.R
\name{combineModalities}
\alias{combineModalities}
\title{combine the UDHS profile of modalities}
\usage{
combineModalities(object)
}
\arguments{
\item{object}{a Bart object}
}
\value{
a Bart object
}
\description{
This function combine the UDHS profile of modalities to generate
a bimodal profile
}
