% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_bimodal.R
\name{.aggregate_rank}
\alias{.aggregate_rank}
\title{Aggregate two rank lists}
\usage{
.aggregate_rank(vector_1, vector_2, n_valid, method)
}
\arguments{
\item{vector_1}{a numeric vector, names are entry ids}

\item{vector_2}{a numeric vector, names are entry ids}

\item{n_valid}{number of top ranks to consider}

\item{method}{aggregation method}
}
\description{
Aggregate two rank lists
}
