% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{get_library}
\alias{get_library}
\title{Download bart2 library}
\usage{
get_library(lib_dir)
}
\arguments{
\item{lib_dir}{path to store library data}
}
\value{
path where library date is stored
}
\description{
Download bart2 library
}
\examples{
get_library(path_of_library)
}
