% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{install_bart2}
\alias{install_bart2}
\title{install bart2 python package}
\usage{
install_bart2(bart_dir = NULL, lib = NULL)
}
\arguments{
\item{bart_dir}{path to install bart2}

\item{lib}{path of bart2 library data}
}
\value{
no return
}
\description{
install bart2 python package
}
